Move CommandParser to its own module
